+++ src/REvoDesign/clients/QtSocketConnector.py
@@ -16,8 +16,8 @@ from PyQt5 import QtCore, QtNetwork, QtWebSockets  # type: ignore
 from REvoDesign import ConfigBus, SingletonAbstract, issues
 from REvoDesign.common.MutantTree import MutantTree
 from REvoDesign.logger import root_logger
-from REvoDesign.tools.client_tools import SSLCertificateManager
 from REvoDesign.tools.customized_widgets import refresh_tree_widget
+from REvoDesign.tools.ssl_certificates import SSLCertificateManager
 from REvoDesign.tools.utils import run_worker_thread_with_progress
 
 logging = root_logger.getChild(__name__)