+++ src/REvoDesign/editor/__init__.py
@@ -1,0 +1,5 @@
+from .monaco.monaco import menu_edit_file
+
+__all__ = [
+    'menu_edit_file'
+]