+++ src/REvoDesign/tools/system_tools.py
@@ -8,8 +8,8 @@ from dataclasses import dataclass, field
 
 from immutabledict import immutabledict
 
-from REvoDesign import SingletonAbstract, issues
-
+from .. import issues
+from ..basic import SingletonAbstract
 from .package_manager import issue_collection
 
 
@@ -55,7 +55,7 @@ class SystemInfoReduced(SingletonAbstract):
 
     def __init__(self):
         if not hasattr(self, "initialized"):
-            self.info: immutabledict = immutabledict(issue_collection())
+            self.info: immutabledict = immutabledict(issue_collection(network=False))
             self.initialized = True